requirements.txt
sentence-transformers
streamlit
chromadb
pymupdf
pysqlite3-binary
requests
python-dotenv
protobuf==3.20.3